+++ src/pyrepo/util.py
@@ -13,7 +13,6 @@ import sys
 from textwrap import fill
 import time
 from typing import TYPE_CHECKING, Any, Optional, TextIO
-import click
 from in_place import InPlace
 from intspan import intspan
 from jinja2 import Environment, PackageLoader
@@ -175,34 +174,6 @@ def rewrap(s: str) -> str:
         replace_whitespace=False,
         width=79,
     )
-
-
-def optional(
-    *decls: str, nilstr: bool = False, **attrs: Any
-) -> Callable[[click.decorators.FC], click.decorators.FC]:
-    """
-    Like `click.option`, but no value (not even `None`) is passed to the
-    command callback if the user doesn't use the option.  If ``nilstr`` is
-    true, ``--opt ""`` will be converted to either `None` or (if ``multiple``)
-    ``[]``.
-    """
-
-    def callback(ctx: click.Context, param: click.Parameter, value: Any) -> None:
-        assert param.name is not None
-        if attrs.get("multiple"):
-            if nilstr and value == ("",):
-                ctx.params[param.name] = []
-            elif value != ():
-                ctx.params[param.name] = value
-        else:
-            if nilstr and value == "":
-                ctx.params[param.name] = None
-            elif value is not None:
-                ctx.params[param.name] = value
-
-    if not attrs.get("multiple"):
-        attrs["default"] = None
-    return click.option(*decls, callback=callback, expose_value=False, **attrs)
 
 
 def yield_lines(fp: TextIO) -> Iterator[str]: